Fix --prefer-pkgs for noinstall="1" packages in kiwi builds
For kiwi to find those packages, they have to be in one of the listed repos.
Create a "_local" repo with all prefer-pkgs in buildinfo.

Note: The code currently doesn't look at noinstall at all, so any package
in the buildinfo which is provided by --prefer-pkgs is treated as noinstall="1"
and noinstall="0" simultaneously currently...