Support package linking of packages from scmsync projects
It needs to remove the scmsync tag in package meta in that
case (when creating a new package meta based on source).

I didn't that global in replace_pkg_meta() on purpose since
this should be a case by case decision when it makes sense.
We could add this there via an option though.

There is another problem in the api that it delivers
package meta without title/description elements, but
it is not accepting it. This is submitted there
via a seperate pull request.