Allow `osc rpmlint` to infer prj/pkg from CWD.
Makes `osc rpmlint` infer the project and package when working from a
locally checked out package dir, if two or fewer arguments are passed.
All four arguments are still required when called from a location that
is not an osc package working copy.

This makes the behaviour of `osc rpmlint` similar to `osc buildlogtail`.