Support copy of scmsync packages
There are two ways, either copy the scmsync definition or drop it
when creating a new package meta.

If we keep it, we must not ask OBS to copy sources, it is syncing it
anyway.

We could add another option to skip the scmsync tag copy, but we keep
it as default, because we don't want to give a different view of the
the sources to the user.  The client side copy is doing this.