Assume openSUSE:Factory if project cannot be determined
Don't assume "factory" branch is needed to build against Factory.
This should be the default build configuration.

As-is, it's impossible to initiate a build from other branches,
even when using --alternative-project= so having a default is probably
advisable. Also, devel projects are being imported with "main" as
their default branches.