When doing an mr from a working copy check for unchecked modifications
Maintenance requests may be done from a working copy of a package
to just update a single code stream at once.
In this situation, this working copy may have been used to implement
this change. Thus, like with submit requestes, check if the working
copy has local changes that have not been checked in and warn the
user about it. This should help to prevent resource-intensive accidents.

Signed-off-by: Egbert Eich <eich@suse.com>