osc.spec: add "Provides: git-obs"
Other git plugins, such as LFS, are shipped as "git-foo" packages.
A reasonable user might expect to be able to install the git obs
plugin by installing a "git-obs" package, but this does not work
currently.

With this patch, a user attempting to install "git-obs" will get
the git obs plugin via osc.

Signed-off-by: Nathan Cutler <ncutler@suse.com>