Make `buildlog` work outside of osc package directory
With the git integration there may not always be a checked out package
available. With this change, `osc buildlog` supports to be called with
more than 2 arguments: in this, the first argument is interpreted as
the project, the 2nd as the package while the third denotes the repository
used. Arch may be specified as a 4th argument, if omitted, take the
host arch. This implements issue #1894.

Signed-off-by: Egbert Eich <eich@suse.com>